Conflicts:
	library/SpatialDomains/Conditions.h
	solvers/CompressibleFlowSolver/CMakeLists.txt
	solvers/CompressibleFlowSolver/EquationSystems/CompressibleFlowSystem.cpp
	solvers/CompressibleFlowSolver/EquationSystems/CompressibleFlowSystem.h
	solvers/CompressibleFlowSolver/EquationSystems/EulerADCFE.h
	solvers/CompressibleFlowSolver/EquationSystems/NavierStokesCFE.cpp
	solvers/CompressibleFlowSolver/Tests/Couette_FRHU_LDG_SEM.tst
	solvers/CompressibleFlowSolver/Tests/Couette_WeakDG_LDG_SEM_3DHOMO1D_FFT.tst
	solvers/CompressibleFlowSolver/Tests/CylinderSubsonic_NS_WeakDG_LDG_GAUSS.tst
	solvers/CompressibleFlowSolver/Tests/CylinderSubsonic_NS_WeakDG_LDG_SEM.tst
	solvers/CompressibleFlowSolver/Tests/CylinderSubsonic_NS_WeakDG_LDG_SEM_VariableMu_par.tst
	solvers/CompressibleFlowSolver/Tests/CylinderSubsonic_NS_WeakDG_LDG_variableMu.tst

+++ library/LibUtilities/BasicUtils/FieldIO.cpp
@@ -432,23 +432,28 @@ namespace Nektar
                 else // only load relevant partitions
                 {
                     int i,j;
-                    map<int,int> FileIDs; 
+                    map<int,vector<int> > FileIDs;
+                    map<int,vector<int> >::iterator it;
                     set<int> LoadFile;
-                    
+
                     for(i = 0; i < elementIDs_OnPartitions.size(); ++i)
                     {
                         for(j = 0; j < elementIDs_OnPartitions[i].size(); ++j)
                         {
-                            FileIDs[elementIDs_OnPartitions[i][j]] = i;
+                            FileIDs[elementIDs_OnPartitions[i][j]].push_back(i);
                         }
                     }
                     
                     for(i = 0; i < ElementIDs.num_elements(); ++i)
                     {
-                        ASSERTL1(FileIDs.count(ElementIDs[i]) != 0,
-                                 "ElementIDs  not found in partitions");
-                        
-                        LoadFile.insert(FileIDs[ElementIDs[i]]);
+                        it = FileIDs.find(ElementIDs[i]);
+                        if (it != FileIDs.end())
+                        {
+                            for (j = 0; j < it->second.size(); ++j)
+                            {
+                                LoadFile.insert(it->second[j]);
+                            }
+                        }
                     }
                     
                     set<int>::iterator iter; 
@@ -518,20 +523,19 @@ namespace Nektar
 
             for (int t = 0; t < fileNames.size(); ++t)
             {
+                if(elementList[t].size())
+                {
+                    TiXmlElement * elemIDs = new TiXmlElement("Partition");
+                    root->LinkEndChild(elemIDs);
+                    
+                    elemIDs->SetAttribute("FileName",fileNames[t]);
+                    
+                    string IDstring;
+                    
+                    GenerateSeqString(elementList[t],IDstring);
 
-                ASSERTL1(elementList[t].size() > 0,
-                         "Element list must contain at least one value.");
-
-                TiXmlElement * elemIDs = new TiXmlElement("Partition");
-                root->LinkEndChild(elemIDs);
-
-                elemIDs->SetAttribute("FileName",fileNames[t]);
-
-                string IDstring;
-
-                GenerateSeqString(elementList[t],IDstring);
-
-                elemIDs->LinkEndChild(new TiXmlText(IDstring));
+                    elemIDs->LinkEndChild(new TiXmlText(IDstring));
+                }
             }
 
             doc.SaveFile(outFile);
@@ -1116,14 +1120,17 @@ namespace Nektar
             fs::path specPath (outname);
 
             // Remove any existing file which is in the way
-            try
+            if(m_comm->RemoveExistingFiles())
             {
-                fs::remove_all(specPath);
-            }
-            catch (fs::filesystem_error& e)
-            {
-                ASSERTL0(e.code().value() == berrc::no_such_file_or_directory,
-                         "Filesystem error: " + string(e.what()));
+                try
+                {
+                    fs::remove_all(specPath);
+                }
+                catch (fs::filesystem_error& e)
+                {
+                    ASSERTL0(e.code().value() == berrc::no_such_file_or_directory,
+                             "Filesystem error: " + string(e.what()));
+                }
             }
 
             // serial processing just add ending.
@@ -1329,7 +1336,12 @@ namespace Nektar
         int FieldIO::CheckFieldDefinition(const FieldDefinitionsSharedPtr &fielddefs)
         {
             int i;
-            ASSERTL0(fielddefs->m_elementIDs.size() > 0, "Fielddefs vector must contain at least one element of data .");
+
+            if(fielddefs->m_elementIDs.size() == 0) // empty partition
+            {
+                return 0;
+            }
+            //ASSERTL0(fielddefs->m_elementIDs.size() > 0, "Fielddefs vector must contain at least one element of data .");
 
             unsigned int numbasis = 0;
 
@@ -1529,7 +1541,5 @@ namespace Nektar
             
             return datasize;
         }
-        
-
     }
 }
